Use network chain height during transaction relay
Use the network chain height to validate spending a coinbase output when
relaying a transaction.

+++ src/main/java/org/ScripterRon/JavaBitcoin/Parameters.java
@@ -161,7 +161,7 @@ public class Parameters {
     public static long txRejected;
 
     /** Network chain height */
-    public static long networkChainHeight;
+    public static int networkChainHeight;
 
     /** List of peer requests that are waiting to be sent */
     public static final List<PeerRequest> pendingRequests = new LinkedList<>();
